requirements.txt
boto3==1.28.38
botocore==1.31.38
jmespath==1.0.1
numpy==1.25.2
pandas==2.1.0
psycopg2-binary==2.9.7
python-dateutil==2.8.2
pytz==2023.3
s3transfer==0.6.2
six==1.16.0
tzdata==2023.3
urllib3==1.26.16
